refactor: 旧 /talk/chat/ エンドポイントを廃止
セッションベースの新エンドポイント（/talk/sessions/.../messages/）が完成した
ため、過去履歴を都度クライアントが送る旧 stateless API を廃止。

削除:
- views.py: TalkChatView クラス全体
- urls.py: path("chat/", ...) ルート
- serializers.py: ChatMessageSerializer / ChatRequestSerializer /
  ChatResponseSerializer / CHAT_MESSAGES_MAX_COUNT 定数

テスト再構成:
- tests/features/talk/test_chat.py を test_synthesize_chat_service.py に
  リネームし、TestTalkServiceChat と TestTalkServiceChatMaxTokens のみ残置
  （TalkService.synthesize_chat は新セッション view からも内部利用するため）
- TestChatRequestSerializer / TestChatResponseSerializer / TestTalkChatView
  は新仕様で意味を失うため削除
- 旧 view にあった throttle (talk_chat) のレート制限テストは
  test_session_message_views.py::test_messages_endpoint_rate_limited として
  新エンドポイントへ移植

検証:
- pytest 全 621 pass（旧 27 ケース削除、新 1 ケース追加で純減 26）
- ruff check / format 全 208 ファイル通過
