+++ django_api/features/talk/serializers.py
@@ -50,70 +50,6 @@ class TalkResponseSerializer(serializers.Serializer):
     )
 
 
-class ChatMessageSerializer(serializers.Serializer):
-    """チャット 1 メッセージ用シリアライザ."""
-
-    role = serializers.ChoiceField(
-        choices=[("user", "user"), ("assistant", "assistant")],
-        help_text="メッセージの役割（user / assistant）",
-    )
-    content = serializers.CharField(
-        max_length=4000,
-        trim_whitespace=False,
-        allow_blank=False,
-        help_text="メッセージ本文（最大 4000 文字）",
-    )
-
-
-CHAT_MESSAGES_MAX_COUNT = 50
-
-
-class ChatRequestSerializer(serializers.Serializer):
-    """チャットリクエスト用シリアライザ."""
-
-    config_name = serializers.CharField(
-        required=True,
-        max_length=50,
-        help_text="設定名（管理画面で登録した name）",
-    )
-    messages = ChatMessageSerializer(
-        many=True,
-        allow_empty=False,
-        help_text=(f"会話履歴（1〜{CHAT_MESSAGES_MAX_COUNT} 件、末尾は role='user'）"),
-    )
-
-    def validate_messages(self, value):
-        if len(value) > CHAT_MESSAGES_MAX_COUNT:
-            raise serializers.ValidationError(
-                f"messages は {CHAT_MESSAGES_MAX_COUNT} 件以下にしてください"
-            )
-        if value[-1]["role"] != "user":
-            raise serializers.ValidationError(
-                "messages の末尾は role='user' にしてください"
-            )
-        return value
-
-
-class ChatResponseSerializer(serializers.Serializer):
-    """チャットレスポンス用シリアライザ."""
-
-    message = ChatMessageSerializer(
-        help_text="生成された assistant メッセージ",
-    )
-    audio_data = serializers.CharField(
-        required=False,
-        allow_null=True,
-        default=None,
-        help_text="Base64 エンコードされた音声データ（TTS 有効時）",
-    )
-    audio_format = serializers.CharField(
-        required=False,
-        allow_null=True,
-        default=None,
-        help_text="音声フォーマット（wav, mp3, ogg）",
-    )
-
-
 class ConfigItemSerializer(serializers.Serializer):
     """設定一覧の各アイテムのシリアライザー."""
